Extract method for setting file caption
