+++ src/flickypedia/apis/__init__.py
@@ -1,0 +1,6 @@
+from .wikimedia import UnknownWikimediaApiException, WikimediaApi
+
+__all__ = [
+    "UnknownWikimediaApiException",
+    "WikimediaApi",
+]
